requirements.txt
jinja2
python-slugify
pyinstaller
